+++ UserManager.cpp
@@ -46,7 +46,7 @@ int UserManager::getNewUserId()
 
 bool UserManager::isLoginExist(string login)
 {
-    for (int i = 0; i <users.size(); i++)
+    for (size_t i = 0; i <users.size(); i++)
     {
         if (users[i].getLogin() == login)
         {
